Time system changed from turns to energy-based queue
Previously, Pantheon played in a series of discrete turns, with one action given to each turn tentatively. This system demanded an unsubtle implementation as players grew powerful, as they would have to find themselves lining up large sequences of actions for each turn. With an energy-based system like those in Crawl or ADOM, an increase in the number of actions which can be performed by the player in the same timeframe does not change the way the player interfaces the gameplay; the only difference is that the world will *appear* to slow down for the player, which may have the positive consequence of a more powerful-feeling progression.

Note that while this commit does encompass this change and all changes needed to keep a working order in the game, the new system itself is not finished. Outstanding issues include:

- Action system is unfinished, with many actions not yet converted to first-class and BaseAction still relying on an actor field (even though many actions are guaranteed to not be tied to any actor).
- Items are unfinished; currently the only thing any Flask can do is heal the player. A new framework is needed for both item instances and flyweights.
- The new turn system is not supported by the UI.
- Documentation is almost completely absent.