Cleaned up variables and properties
This includes adding default values, standardizing names, and privatizing certain fields to be exposed to the Unity editor but not to the rest of the code.

Additionally, a ReadOnly drawer attribute for the editor has been added, courtesy of It3ration and FuzzyLogic on the Unity Answers forum.

Finally, error checking was added to Database getters.