Dread hamster and relevant parts added
This was to test the feasibility of the current time system in relation to multi-turn actions (specifically attacks). The system may require tweaks later, but it is currently acceptable that the player action takes precedence, seeing as how one taps of meaningful enemies shouldn't be happening much anyways.

Also in this commit:

- The dread hamster is native to the Northern Valley, which now has a population set.
- NPC now uses OnEnable() instead of Start() to initialize its spriteRenderer, preventing a NullReferenceException when refreshing the FOV of a newly-generated level (which requires the spriteRenderer before the first frame update).
- Claws and teeth are new body part types: the misc type has been removed for now.