New projectile behaviour
Rather than instantiating a tangible projectile which then acts as one would, LineProjAction handles the projectile's travel and effects invisibly, and then handles particle effects afterwards. The only gameplay difference is that now projectiles have damage ranges, accuracy, and a variable for piercing through actors: most of the reason for this commit is to improve visual polish.

Also, Hits now validate construction arguments.

As a footnote, Spell.cs was added to the project prior to the projectile change. This was the first part of the upcoming implementation of spells, which was going to use Patson's Magic Bullet as a testing ground. Afterwards, the projectile rework followed to prepare projectiles for the new spell, but spells are going to come later, as the current new changes deserve their own commit.