New ECS adopted by items
Items always had a form of ECS (Melee, Ranged, Ammo, etc.) but every item had fields for these components that would simply go null in the absence of one. This commit makes these components implement an interface, and every item has a dictionary of these interfaces keyed by an enum (to prevent multiplicity).

Additionally, an instance of item BaseData no longer has dead objects in it for all the components it does not possess, but rather keeps a list for ScriptableObject wrappers containing these components. This brings it nicely into line with the already-standing system of action wrappers.

There is now a largely-unified system of components which can be applied to anything given the right framework in the entity: an Appendage can now be used as a ranged weapon, for instance.

Other changes in this commit:

- Pantheon.Components namespace.
- WearAction now has handling for all possible EquipType cases.
- FIX: Actors.Equipment was missing a headwear field.
- Moved Melee to Components.cs, Pantheon.Components namespace.