Inventory now reflects starting items
Additionally, the building of the first level is now moved back to Game.Awake() from Start(). This allows the player to be sufficiently initialized to allow the HUD to start, as Game.Start() will always be called afterwards for no discernible reason. This is meant to allow the game to start even if the HUD is enabled (see commit 4d89ac3), but seeing as how a save/load system is next on the roadmap, this change probably makes no difference.