Rolled all item data into single class, now looked up by string
Previously, ItemData was an abstract class with numerous inheritors, each of which held nothing but a different type of enum by which they could be looked up. As Pantheon moves to (possibly) become Reliquary, this is less logical than ever before. This is likely to be the first step in a migration; in theory, since the database is never accessed by index or sorted, it could take the form of a single collection (HashSet?), populated programatically, and accessed via a single dictionary.

- What once was "BaseData" is now known as "Content".
- Some identifiers referring to "BodyPart" are now "Appendage".
- /Pantheon step in CreateAssetMenu.
- ItemWeights has gone out in favour of a ScriptableObject model.
- SerializableRandomPick, for use in the WeightTable SO.
- ItemFactory is gone: now, there is only a static factory method in Item.
- Intro now passes the starting weapon to the Game controller in the form of an ItemData object, tied to the button.
- All enums for referencing items are gone; any classes which kept them in fields now use their data forms instead.
- PantheonEditor.Database. Contains PopulateDatabase(), which can be used to fill Core.Database.itemList with all existing ItemData assets labelled "Item" by the editor.