Returned databasing to list/dict model with lookup by enum
String IDs for content objects aren't safe, and so constants are needed to ease development. Things are likely to stay this way unless/until a new content model without a database can be devised.

What is different now is that the database can now be automatically populated from the editor based on the Assets directory, speeding the workflow of adding new content.

Additionally:

- Database list and dictionary capacities are now set based on the number of assets in the project.
- Any classes which were named "{Category}Data" are now "{Category}Def" (mostly because TerrainData is already a type in Unity).
- All enums used to key content objects are known as "{Category}ID".
- Content object enums now have 0 more appropriately named "Default" instead of "None".
- Finished refactoring "Body Part" to "Appendage".
- Cleaned up some def class default values, properties (removed private setters).
- PantheonEditor.Database.PopulateDatabase() now saves the Game scene following operation.
- More database accessors now report the name of an illegal argument if one is given.