Started migrating Pantheon to new ECS
While the design direction of the game is being decided, it is best that a new system be put in place to make the game more flexible in adapting to whatever design comes. The new ECS will replace actors, items, features, terrain, and likely more with a more powerful system of entities and components, with some systems to handle updates which are needed on a regular basis.

- Started using AssetBundles. The existing database isn't going anywhere, but the game will not use it for the foreseeable future as an extended test for feasibility.
- BaseAction has been renamed to Command. Eventually all actions will be refactored like this (the pattern at play is named "command" in the first place).
- CameraControl script to allow dragging the camera around with the right mouse.
- Actually fixed the diagonal movement keys.