Freed entity-components from Unity (...again?)
Upon further review, Pantheon's entity components are incompatible with Unity's MonoBehaviour system because many entities can't or shouldn't be GameObjects (e.g. items). As such, GameObjects are back to being a presentation layer again, but now with an added touch: GameObjects representing entities have a container MonoBehaviour which stores a pointer to that entity. This should make interactions between the engine and the underlying entities much easier.

Additionally:
- JSON serialization is working again.
- BaseComponent has been renamed to EntityComponent.
- Editor window for designing templates.
- Fixed actor command field not being nullified following execution.
- Large-scale cleanup of Assets folder.