Foundations for melee combat implemented
Currently, much of the functionality needed to bring this feature together are missing. Logging needs to be made global somehow and entities can't currently be properly destroyed/killed because of the inaccesibility of the turn scheduler. For now, however, the core logic of dealing damage to a target is present.