Player now spawns in random corner of the map
This is leading up to placing an exit to the level on the opposite corner, and thus getting the player moving between levels.

Work is in progress to put the player in a vault to start with, but because a much larger system will have to built to ensure that the vault has open clearance in the terrain, this will suffice.

- Already-filled check moved after all other checks in flood fillers.