World generation system revised
The world layout is now specified by defining builders, both of which are JSON files and flyweights now. A builder is a unique specification for a level in the world, and knows its layout, connections, and NPC populations. The IDs of these builders are then put into a dictionary known as the world plan and invoked to create levels upon request in the same way as before.