Removed lights from EntityGameObjects
When many of these lights are instantiated, they cause an unacceptable performance drop even when disabled. If the game is going to keep 2D lighting, these lights will eventually have to be instantiated and destroyed as entities come into the FOV, but since they don't currently add much to the game's feel, I'm just going to scrap them for now.