AI memory, new turn scheduler handling
AI now has a length of time for which they will remember the player while not being in sight, after which they will stop caring and go back to a state of rest. This makes it possible to keep most actors in the world out of the turn scheduler when they were already doing nothing, giving a massive performance savings.