Wolverson; a grenade-slinging boss-like enemy
This commit decouples talents from entities. Moving forward, the goal is that a talent is something learned by a character and which makes use of that character's weapons and body parts based on its specific needs. This makes the system more like Path of Exile's and less like Terraria's, to provide a frame of reference.

The reasoning for this change in paradigm is to facilitate talents which can use multiple weapons or body parts. It also helps prevent repeating talents which do the same thing over the multiple items (e.g. "throw hatchet" and "throw knife" are virtually identical).

- Fix template deserialization comparing wield array sizes incorrectly.
- Entity.Move() now only involves Level.MoveEntity() if entity is "on the ground"; that is, not in an inventory or wielded.
- Consider argument array length instead of instance array length in Wield.ForceWield().
- TemplateConverter, currently unused.
- TossTalent names it projectile properly.
- IList<T>.Shuffle() extension.
- Cleaned up DialogueHelpers.
- Use ForceWield() in Entity(EntityTemplate).